End session after 24 hours and redirect to home page

+++ src/actions/cars/DeleteCar.ts
@@ -3,11 +3,16 @@
 import { getServerSession } from "next-auth";
 import { authOptions } from "@/auth/auth";
 import DatalogViewerSession from "@/interfaces/DatalogViewerSession";
+import { redirect } from "next/navigation";
 
 const datalogApiUrl = process.env.DATALOG_API_URL;
 
 export default async function DeleteCar(carId: number) {
   const session = (await getServerSession(authOptions)) as DatalogViewerSession;
+
+  if (session === null) {
+    redirect("/");
+  }
 
   const response = await fetch(`${datalogApiUrl}/api/cars/${carId}`, {
     method: "DELETE",
