+++ src/actions/tracks/DeleteTrack.ts
@@ -1,14 +1,18 @@
 "use server";
 
-import Track from "@/interfaces/Track";
 import { getServerSession } from "next-auth";
 import { authOptions } from "@/auth/auth";
 import DatalogViewerSession from "@/interfaces/DatalogViewerSession";
+import { redirect } from "next/navigation";
 
 const datalogApiUrl = process.env.DATALOG_API_URL;
 
 export default async function DeleteTrack(trackId: number) {
   const session = (await getServerSession(authOptions)) as DatalogViewerSession;
+
+  if (session === null) {
+    redirect("/");
+  }
 
   const response = await fetch(`${datalogApiUrl}/api/tracks/${trackId}`, {
     method: "DELETE",
